[Xposed] Disable direct branching optimization
Under certain circumstances (e.g. the method must be part of the boot image and
cannot be a virtual method), ART calls methods directly using their known code
offset in the image. That makes hooking these methods impossible (or rather
ineffective, as the hook isn't triggered). Hence, this optimization has to be
disabled.