[Xposed] Disable accessing TypedArray subclass fields via offset
Xposed modifies the size of TypedArray objects to match those of XTypedArray
objects in order to change the class of created objects later. This leads to
issues when other apps create their own subclass of TypedArray with additional
fields. When compiling the app, the compiler doesn't know that the TypedArray
object will be a few bytes bigger, so the offsets of the additional fields
differ between runtime and compile time. If the compiler optimizes the access
to these fields by addressing them via their offsets, it will lead to crashes.

Therefore, we can't allow this kind of optimization for subclasses of TypedArray.