[Xposed] Add method modifier to ignore AOT code
Methods with this modifier set shouldn't use their AOT code, most likely
because one of the methods they call has been hooked. Therefore prevent
methods from calling or resetting the entry points to the AOT code and
use the interpreter instead. Also triggers deoptimization when returning
to such a method.

This still allows the method to be compiled with JIT, which will
consider the new circumstances and avoid optimizations that are now
impossible.

Changing the existing entry points will be part of a later change.