[Xposed] Invalidate the callers of hooked methods
This puts the pieces of the previous commits together. It set the
IgnoreAotCode modifier, invalidates JIT code and enables instrumentation
of the current stack frames (which will cause deoptimization when
returning to the new invalidated callers of the hooked method).

Besides that, it will also replace pointers to the method with the
original method in the stack and in the JIT code cache.